requirements.txt
flask-socketio
gunicorn
eventlet
pyenchant~=3.2.0